+++ rai/client.go
@@ -35,7 +35,7 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
-const userAgent = "raictl/" + Version
+const userAgent = "rai-sdk-go/" + Version
 
 type PreRequestHook func(*http.Request) *http.Request
 
@@ -1639,7 +1639,7 @@ func (c *Client) CreateSnowflakeIntegration(
 }
 
 func (c *Client) DeleteSnowflakeIntegration(name string, adminCreds *SnowflakeCredentials) error {
-	req := deleteIntegrationRequest{}
+	req := deleteSnowflakeIntegrationRequest{}
 	req.Snowflake.Admin = *adminCreds
 	return c.Delete(makePath(PathIntegrations, name), nil, &req, nil)
 }
@@ -1655,6 +1655,60 @@ func (c *Client) GetSnowflakeIntegration(name string) (*Integration, error) {
 func (c *Client) ListSnowflakeIntegrations() ([]Integration, error) {
 	var result []Integration
 	if err := c.Get(PathIntegrations, nil, nil, &result); err != nil {
+		return nil, err
+	}
+	return result, nil
+}
+
+//
+// Snowflake Database Links
+//
+
+func (c *Client) CreateSnowflakeDatabaseLink(
+	integration, database, schema, role string, creds *SnowflakeCredentials,
+) (*SnowflakeDatabaseLink, error) {
+	var result SnowflakeDatabaseLink
+	path := makePath(PathIntegrations, integration, "database-links")
+	req := createSnowflakeDatabaseLinkRequest{}
+	req.Snowflake.Database = database
+	req.Snowflake.Schema = schema
+	req.Snowflake.Role = role
+	req.Snowflake.Credentials = *creds
+	if err := c.Post(path, nil, &req, &result); err != nil {
+		return nil, err
+	}
+	return &result, nil
+}
+
+func (c *Client) DeleteSnowflakeDatabaseLink(
+	integration, database, schema, role string, creds *SnowflakeCredentials,
+) error {
+	name := fmt.Sprintf("%s.%s", database, schema)
+	path := makePath(PathIntegrations, integration, "database-links", name)
+	req := deleteSnowflakeDatabaseLinkRequest{}
+	req.Snowflake.Role = role
+	req.Snowflake.Credentials = *creds
+	return c.Delete(path, nil, &req, nil)
+}
+
+func (c *Client) GetSnowflakeDatabaseLink(
+	integration, database, schema string,
+) (*SnowflakeDatabaseLink, error) {
+	var result SnowflakeDatabaseLink
+	name := fmt.Sprintf("%s.%s", database, schema)
+	path := makePath(PathIntegrations, integration, "database-links", name)
+	if err := c.Get(path, nil, nil, &result); err != nil {
+		return nil, err
+	}
+	return &result, nil
+}
+
+func (c *Client) ListSnowflakeDatabaseLinks(
+	integration string,
+) ([]SnowflakeDatabaseLink, error) {
+	var result []SnowflakeDatabaseLink
+	path := makePath(PathIntegrations, integration, "database-links")
+	if err := c.Get(path, nil, nil, &result); err != nil {
 		return nil, err
 	}
 	return result, nil